requirements.txt
annotated-types==0.7.0
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
blis==0.7.11
bs4==0.0.2
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
colorama==0.4.6
comtypes==1.4.2
confection==0.1.4
cymem==2.0.8
filelock==3.14.0
Flask==3.0.3
fsspec==2024.5.0
h11==0.14.0
idna==3.7
intel-openmp==2021.4.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
langcodes==3.4.0
language_data==1.2.0
marisa-trie==1.1.1
MarkupSafe==2.1.5
mkl==2021.4.0
MouseInfo==0.1.3
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.3
nltk==3.8.1
numpy==1.26.4
outcome==1.3.0.post0
packaging==24.0
pillow==10.3.0
preshed==3.0.9
pt-core-news-lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.7.0/pt_core_news_lg-3.7.0-py3-none-any.whl#sha256=4eed7801647c39ced078eb325d0ff8eb2ea5d4945501b8ee898dced85c686717
PyAudio==0.2.14
PyAutoGUI==0.9.54
pycparser==2.22
pydantic==2.7.1
pydantic_core==2.18.2
PyGetWindow==0.0.9
PyMsgBox==1.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
PyScreeze==0.1.30
PySocks==1.7.1
pyttsx3==2.90
pytweening==1.2.0
pywhatkit==5.4
pywin32==306
regex==2024.5.15
requests==2.32.2
selenium==4.21.0
setuptools==70.0.0
smart-open==6.4.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.5
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SpeechRecognition==3.10.4
speedtest-cli==2.1.3
srsly==2.4.8
sympy==1.12
tbb==2021.12.0
thinc==8.2.3
torch==2.3.0
tqdm==4.66.4
trio==0.25.1
trio-websocket==0.11.1
typer==0.9.4
typing_extensions==4.12.0
urllib3==2.2.1
wasabi==1.1.2
weasel==0.3.4
Werkzeug==3.0.3
wikipedia==1.4.0
wsproto==1.2.0
